refactor(ui): extract view-model/peer-status.ts — derivePeerUiStatus + derivePeerTrustSummary
Move the two peer status derivations into view-model/peer-status.ts:
derivePeerUiStatus (connected / needs-repair / offline / waiting) and
derivePeerTrustSummary (the longer trust blurb). The barrel re-exports
both so sync-peers.tsx and other call sites stay on the same import
path.
